pyproject.toml
[project]
name = "hermes-station"
version = "0.0.1"
description = "Single-container deployment scaffolding for Hermes Agent + WebUI."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
authors = [{name = "Rory Ford", email = "me@roryford.com"}]

dependencies = [
    "starlette>=0.40",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "python-multipart>=0.0.20",
    "itsdangerous>=2.2",
]

[project.optional-dependencies]
# Pinned upstream — tracked by Renovate's pep621 manager.
# hermes-webui is pinned in the Dockerfile (regex manager) because it has no pyproject.toml upstream.
# Kept as an extra so local dev / lint / unit-test cycles don't pull the full agent dep tree.
hermes = [
    "hermes-agent[messaging] @ git+https://github.com/NousResearch/hermes-agent.git@v2026.5.7",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
# Required so the `hermes` extra can reference hermes-agent by git URL.
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["hermes_station"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
